# frontend/src/api/generated/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CreateTimelineRevisionRequest": {
        "properties": {
          "derived_from": {
            "type": "string"
          },
          "episode": {
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "TimelineItem"
            },
            "type": "array"
          },
          "project": {
            "type": "string"
          },
          "subtitle_revision": {
            "type": "string"
          }
        },
        "required": [
          "project",
          "items"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "description": "稳定错误码，代码里以此分支，不要解析文案",
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "success": {
            "const": false,
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error",
          "code"
        ],
        "type": "object"
      },
...
      "TimelineRevisionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "TimelineRevision"
            },
            "type": "array"
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "total": {
            "type": "integer"
          }
        },
        "required": [
          "success",
          "items"
        ],
        "type": "object"
      },
      "TimelineRevisionResponse": {
        "properties": {
          "item": {
            "$ref": "TimelineRevision"
          },
          "success": {
            "const": true,
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "本规格是**前端与后端的唯一契约**。前端客户端由 `scripts/generate_client.py` 从本文件生成，`scripts/check_generated_client.py` 在 CI 校验生成物是否过期。\n\n**手工改这个文件没有意义**：它是生成物，改了会被下一次生成覆盖。",
    "title": "漫剧锻造台 API",
    "version": "2026-10-07.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/timeline/fingerprints/{fingerprint}": {
      "get": {
        "operationId": "getTimelineByFingerprint",
        "parameters": [
          {
            "description": "合成指纹",
            "in": "path",
            "name": "fingerprint",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "无匹配 revision"
          }
        },
        "summary": "按合成指纹反查 revision",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions": {
      "get": {
        "operationId": "listTimelineRevisions",
        "parameters": [
          {
            "description": "按项目过滤",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionListResponse"
                }
              }
            },
            "description": "成功"
          }
        },
        "summary": "时间线 revision 列表（不可变）",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      },
      "post": {
        "operationId": "createTimelineRevision",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTimelineRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法 / 静音未显式声明"
          }
        },
        "summary": "冻结一个时间线 revision",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/latest": {
      "get": {
        "operationId": "getLatestTimelineRevision",
        "parameters": [
          {
            "description": "项目 key",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "集号",
            "in": "query",
            "name": "episode",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "该集尚无时间线"
          }
        },
        "summary": "该集的最新时间线 revision",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/{revision_id}": {
      "get": {
        "operationId": "getTimelineRevision",
        "parameters": [
          {
            "description": "时间线 revision id",
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "revision 不存在"
          }
        },
        "summary": "时间线 revision 详情",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/{revision_id}/derive": {
      "post": {
        "operationId": "deriveTimelineRevision",
        "parameters": [
          {
            "description": "时间线 revision id",
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineRevisionResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "revision 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "无实质变更，不予派生"
          }
        },
        "summary": "由既有 revision 派生新版本（不可原地改）",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/{revision_id}/manifest": {
      "get": {
        "operationId": "getTimelineManifest",
        "parameters": [
          {
            "description": "时间线 revision id",
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderManifestResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "revision 不存在"
          }
        },
        "summary": "渲染清单（合法静音必须前置声明）",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/{revision_id}/preflight": {
      "get": {
        "operationId": "preflightTimelineRevision",
        "parameters": [
          {
            "description": "时间线 revision id",
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            },
            "description": "成功"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "revision 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "存在未声明的合法静音（undeclared_silence）"
          }
        },
        "summary": "合成预检（含 compose_fingerprint）",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
    "/api/timeline/revisions/{revision_id}/render": {
      "post": {
        "operationId": "renderTimelineRevision",
        "parameters": [
          {
            "description": "时间线 revision id",
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RenderVersionResponse"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "参数非法"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "授权者非人工主体"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "revision 不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "预检未通过（存在未声明静音 / 引用已失效版本）"
          }
        },
        "summary": "按**冻结计划**渲染（只消费 plan，缺任何一镜即 fail-closed）",
        "tags": [
          "timeline"
        ],
        "x-contract-source": "manual"
      }
    },
...
  },
  "servers": [
    {
      "description": "同源部署",
      "url": "/"
    }
  ],
...
}